Use place name list and jaro-winkler to fix place names on extraction
A list of place names from database was added to the support_datasheets.
This is a list of place names from db, with unique stems. It can be used to fix small
errors in place names either by finding directly place name with same stem from
list or by using Jaro-Winkler distance to find close match of stemmed place name.
This should alleviate the problem of having conjugated place names in the extraction results
as well as slightly decrease the amount of place names with typos. Still rather rough and
inprecise approach to the problem, but a step forward.