Define path for extraction output results
Now it is possible to optionally configure output path for extractor. For example
one can add path='primaryPerson' to extractor's configure call to force output of the
said extractor to be put under dict key 'primaryPerson'. This structure will be
preserved in json-output. If the path is not defined, extraction results will be
placed directly to the extraction results dict which is nice in sub pipelines.