Generate independent ids for children and spouses
Problem was that before the id was incremented for every person type. This would
be extremely bad thing if the count of children and spouses would change (and they
will change!). Now main id is based on data entries which should be reasonably stable and children and spouses will get incrementing id which is under the
main id.