Finish implementing the simple farm area extractor
Cover some more different cases of how the data is presented. Now
the total area is extracted if it is in ares in the data. Discard
the area if it is in square meters, or if the numbers that make
up the area have a whitespace character between them, as it is too
ambiguous and there is no logic to discern whether it should be a
comma or a dot between the numbers. Also wrote tests for the new
cases and verified that they pass.