Add an extractor for injured in war flag
Added an extractor that checks the data for keywords that indicate
the person in question was injured during a war. The flag is True
or False for males, and should always be null for females. Added
tests for the flag extractor and the extractor is present in the
extraction pipeline.