Add lotta flag extractor to war data extractor
Added the lotta flag extractor to the subpipeline of the war data
extractor and wrote tests to cover the addition. Had to move war data
extractor in the main extraction pipeline because lotta extractor
depends on the results of the spouse extractor for now. Because there
is now a bunch of post-processing hacking in the war data extractor,
made a separate post processing function that goes through the data
and applies post-processing as needed. Added a flag that allows
exclusion of specific extractors under the WarData extractor.