Make dependencies fill in reverse
Changed dependencies to fill in reverse, as this will allow the user to more
intuitively configure contexts - i.e., first start by entering the contexts
for the subclass they are configuring in configure_extractor in the order the
dependencies are declared in in the class, and then moving on to the
subclass's superclass and filling the contexts in similarly.

Also made _deps an empty object instead of None in BaseExtractor's constructor.