Finish making the WarData extractor use the dependency system
Removed the post-process nullifying hack from WarData, as the
extractors themselves now take care of returning None, True or
False depending on the gender of the person we are processing and
whether this is a spouse or a primary person. Also wrote new tests
to test the changed functionality of these extractors.