Improve the regex pattern used for wedding year extraction
The original pattern did not allow for typos in the word, which
there are some cases of, because of the OCR process. Sometimes
'avioit' is also followed by a colon instead of a dot. Also added
detection of OCR typo-variants of dots and colons after 'avioit'.

These changes should give us approximately 1000 more wedding year
hits.