Add test geo db fixture
Creates once per session a copy of geo db so that it is usable by fixture in tests
which need to manipulate the geo database. Note that because the scope is set
as session, the changes persist in between the tests. This is ok for now, but in
future we might want to set autouse=False and scope to function so that
db would be reset on each test.

However, the said behavior *might* slow down the tests so didn't implement it
now. Another option would be to use in-memory copy of the geo db, but
decided to leave it out for now (see the explanation and SO link in the code)