Set migration record region from db if possible
If there is region available in geodb for certain place during migration record
extraction, use it instead of relying to books' listings. These region listings should
be more reliable. However, if there is no information in database, deduce the region
from source text.

This way if coordinates are not available (place not in the geodb) we will at least
set the region to get a rough idea where the place probably is.