Change name detection regular expression pattern
The new pattern allows for detection of names that are not all
upper case, which there are a number of. This change decreases
the number of people whose entries start in the middle of another
person's entry and thus all of their data gets put in that person's
entry. In the first book, this works out to be nine new entries,
and nine entries where the data should be more reliable.