Add command for chunking multiple files & filtering
Had to add a dirty hack fix to the other chunkers to keep them
functional. Those chunkers foreseeably only work on one file ever,
so perhaps this is fine, at least till the chunkers are possibly
merged into one chunker.

Also added variables to the chunk and extract functions in main.py
to deobfuscate the names of the arguments used in those functions.