Import ssdeep only if it is available
Python ssdeep package seemed to be pain to install on OSX.
Didn't bother to figure out if it is even possible without
unreasonable effort so instead made the software and tests
work without it.

ssdeep is only required for chunking tasks, so everything else
still works ok on platforms without the package. ssdeep tests
are now skipped if the package is not available and all other
code should run ok as long as one doesn't try to run the chunking
code.