Remove old dependency resolving system
Old resolving system mostly removed and WarDataExtractor used
as proof of concept for its basic functionality.

Only some of the tests are on since most of the application
and tests are completely broken at the moment. Will fix them
in later commits.

Now the dependencies are resolved by using the ExtractorResultsMap
and YAML defined relationships. Id of the extractor is saved
to _required_dependencies in BaseExtractor's init, and is then
used to fetch the correct data during the extraction.

This however has some problems:
1) The resulting dict structure is a bit weird as can be seen
in WarDataExtractor's sub extractors: self._deps['name']['name']['gender']

2) There is possibility that during resolving, the extractors'
names could collide. For example, if extractor depended on two
NameExtractors. Both have the same name and latter one would
overwrite the former during resolving. We need a way to resolve
this kind of dependencies in a clear manner since now we don't
track contexts and relations (parent. etc) which uniquely
identified the dependencies in the past.