Resolve dependencies to list instead of dict
Decided that for now resolving dependencies to a list is probably eay to go
instead of dict. With dict we would need a unique key for each dependency which
would lead eventually to conflicts if the extractor had two dependencies with same
name.

Solved this now by using list and extractors using their resolved dependencies
with list indexing. While not quite as easy to read from code, this should prevent
collisions based on names. Next we probably want to formally specify the
dependencies of the extractor in code.