Require extractors to declare their requirements in constructor
Each extractor which is dependent on something else needs to specify these
dependencies in their constructor by giving a list of names where the
dependencies will be assigned. For example ['gender'] would mean that
the extractor is dependent on one other extractor, and it will be found during
extractor from self._deps['gender'].

If there is mismatch between amount of requirements provided in YAML and
the amount of the expected dependencies of extractor, an error is raised.