Convert Spouse extractor to use YAML configuration
The subpipeline is now configured using YAML. This caused some
changes on how to test the extractor in an efficient manner.

BaseExtractor can now set its required dependencies by arbitrary string
which will act as id in the ExtractionResultsMap. This lets us
to mock the dependency values in tests directly to the map which
are then resolved to the tested extractor.