Fix broken WarData extraction
Removed the extra pipeline from SpouseExtractor
which broke WarData extraction. Reason being that
other extractors in the pipeline used a substring of the person text, while WarData extractors
expected the full text.

Added a new full_text property to the entry object
so that we can always access the full text if needed.