Fix typos and how extract task help is displayed
The bookpath switch is not actually optional. Technically speaking
the testset switch is optional, but regardless of whether it is
specified or not, the information is always there, as it acts like a
'store true' switch. Neglecting to specify it is the same as
specifying false. Also, marking it as optional in the task creation
forced the invoke task's help information to indicate that the task
expects the testset switch to be accompanied by a string of data,
which was not true.