Fix birthdate extraction logic
Sometimes the location extraction failed in the new implementation. The reason
was incorrect cursor_location passed from BirthDateExtractor. The cursor_location
was incorrectly reset to 0 if date was not found in _find_date() method.

Also removed cursor_location from arguments of DateFinder. The DateExtractor 
never required it (it was always set to 0) and now it would always be 0 so better
to remove it alltogether

.