Fix cursor location position of BirthdayExtractor
Location names were not extracted as well as before because the cursor location
from Birthday extractor was a bit off. It is now set in a similar way as it was before
DateExtractor refactor.

This was fun to debug and quite difficult to test in a meaningful way. Practically
these kinds of changes are only detectable by diffing result json to previous
results. Could we device some better way to extract the meaningful data from
the beginning part of the text? Slight errors in cursor location may change the
results so the current system is quite fragile...