Load bookseries pipeline in a generic BookSeries class
This class will replace the existing "main.py" files in each book series we have.
The idea is that this class gets references to the book series implementation
from yaml-manifest and then serves as an interface for running the book series
related tasks such as extraction and chunking.

Implemented the xml-to-dict conversion here already. It will be removed from our
main main soon.