Chunk Suomen Rintamamiehet from html to xml format
A minimal implementation. Parses the html and splits the persons by
paragraphs which begin with at least three uppercase letters. This will
likely produce at least some incorrect results but that is fine for
now since this is a proof of concept of adding a new book series to the
Kaira. If more accuracy is needed one should improve the chunker logic
by writing a bit more sophisticated heuristics.